Crate warp_backend

WARP – Workflow-aware Advance Resource Planner, is a component of ACROSS orchestration solution. WARP is in charge of manage workflows in form of CWL files, allocating, if needed, proper amount of HPC resources in a specific time frame. WARP leverages on custom SLURM plugin to perform such reservation.

Modules

db_lib 🔒

This module contains every function/method that deals with the database. WARP leverages on Diesel Crate to manage DB interactions, and supports SQLite, MySQL, postgres.

http_server 🔒

Basic HTTP server that provides several APIs to enable a simple communication with workflow manager tools, such as STREAMFLOW https://streamflow.di.unito.it/

reservation_manager 🔒

According to the WARP design, the reservation manager is a separate functions that runs in parallel with the other parts of the WARP and it’s in charge to performe reservations starting from the steps that are in a WAITING state. This module provides two alternatives: i) _check_for_unreserved_steps(), that performs a simple reservation logic allocating resources when the step is ready, and ii) optimization_lookup(), which implements the complex reservation mechanism that will optimize the scheduling of jobs.

slurm_lib 🔒

This module handles interactions with slurm, submitting commands such as scontrol, sresv, etc.

yaml_handle_cmd 🔒

This module provides structs and functions to handle commands received by WARP client through 0MQ

yaml_prs_config 🔒

This module is in charge of parsing configuration file in YAML format in order to complete workflow description retrieved by the CWL submitted. WARP leverages on same configuration file needed by STREAMFLOW tool and follows the same format requirements.

yaml_prs_cwl 🔒

This module provides parsing functionality for CWL workflow description files. This parser doesn’t cover all CWL features, but support the main fields to describe a complete workflow. This parser will be extended to support more advanced features (e.g. loop)

Functions

main 🔒