Module warp_backend::reservation_manager

According to the WARP design, the reservation manager is a separate functions that runs in parallel with the other parts of the WARP and it’s in charge to performe reservations starting from the steps that are in a WAITING state. This module provides two alternatives: i) _check_for_unreserved_steps(), that performs a simple reservation logic allocating resources when the step is ready, and ii) optimization_lookup(), which implements the complex reservation mechanism that will optimize the scheduling of jobs.

Functions

_check_for_unreserved_steps 🔒

Simple reservation logic that looks for steps with satisfied requirements and allocates resources through reservation mechanism (slurm plugin). This method should be only for testing purposes as it bypasses optimization module

optimization_lookup 🔒

This function is the entry point for the Optimizer

reservation_loop

reservation_loop() starts a infinite loop that every $RESERVATION_LOOP_INTERVAL seconds perform a reservation cycle. This means that reservations are performed asynchronously to the submission of workflows