Module warp_backend::yaml_prs_cwl

This module provides parsing functionality for CWL workflow description files. This parser doesn’t cover all CWL features, but support the main fields to describe a complete workflow. This parser will be extended to support more advanced features (e.g. loop)

Structs

NodePayload

Constants

CANCELLED 🔒

COMPLETED 🔒

Job completed and notified by Streamflow

FAILED 🔒

FIREABLE 🔒

Job with a reservation assigned that can be submitted by Streamflow

RUNNING 🔒

Job running

SKIPPED 🔒

WAITING 🔒

Job waiting for the allocation of a specific reservation

Functions

add_nodes 🔒

find_node_rec 🔒

parse_step_dependencies 🔒

parse_yaml_cwl